+++ packages/celest_core/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.2.2-wip
 
-- Adds `Keychain` interface for secure storage of sensitive data.
+- Adds `SecureStorage` interface for storage of sensitive data in the platform keychain.
 
 ## 0.2.1